refactor: change folder api to service and add ts
